Move Transactions class to Models folder and add summary
